+++ src/special/mod.rs
@@ -60,11 +60,12 @@ pub(crate) fn eval_pol(c: &[f64], x: f64) -> f64 {
     acc
 }
 
-pub use beta::{BetaIncError, beta, beta_inc, beta_log, try_beta_inc};
+pub use beta::{beta, beta_inc, beta_log, try_beta_inc, BetaIncError};
 pub use erf::{error_f, error_fc, error_fc_scaled};
 pub use gamma::{
-    GammaDomainError, GammaIncError, GammaIncInvError, PsiError, gamma, gamma_inc, gamma_inc_inv,
-    gamma_log, psi, try_gamma, try_gamma_inc, try_gamma_inc_inv, try_psi,
+    gamma, gamma_inc, gamma_inc_inv, gamma_inc_with_acc, gamma_log, psi, try_gamma, try_gamma_inc,
+    try_gamma_inc_inv, try_gamma_inc_with_acc, try_psi, GammaDomainError, GammaIncAcc,
+    GammaIncError, GammaIncInvError, PsiError,
 };
 pub use normal::{cumnor, dinvnr, dlanor};
 pub use students_t::dt1;